requirements.txt
pynacl>=1.5.0
langchain-core>=0.1.0
pytest>=8.0.0
